+++ changelog.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Use natural sort for alphabetic sorting for files in the list
+- Files without title will now be visible with their filename
 
 ## [3.6.3] - 02.01.2025